Split event logger service into it's own file and pass the client object instead of istantiating the logger in the main program file
